requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
psycopg[binary]
alembic
pydantic>=2
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
tenacity
python-multipart
email-validator
